sprintf/snprintf: Return "more" correct final string length
This change fixes the sprintf/snprintf return value
which is supposed to be number of characters of the
created string.

This change will return the size of the resulting string
up to the buffer length which still isn't up to spec
(in which case it should return string length as it would fit
in the buffer).